pyproject.toml
[project]
name = "chat-claude-bot"
version = "0.1.0"
description = "Discord/Chatwork bot for Claude Code remote control"
requires-python = ">=3.11"
dependencies = [
    "discord-py>=2.3.0",
    "aiohttp>=3.9.0",
    "keyring>=25.0.0",
]

[project.scripts]
chat-claude-bot = "chat_claude_bot.main:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
